Issue #SB-13080 feat: Progress issue resolved

+++ platform-modules/batch-models/src/main/scala/org/ekstep/analytics/job/CourseMetricsJob.scala
@@ -216,7 +216,7 @@ object CourseMetricsJob extends optional.Application with IJob with ReportGenera
         when(expr("progress/leafnodescount * 100").isNull, 100.0)
           .when(expr("progress/leafnodescount * 100") > 100, 100.0)
           .otherwise(expr("progress/leafnodescount * 100")), 2)
-        .cast("double"))
+        .cast("int"))
       .withColumn("generatedOn", date_format(from_utc_timestamp(current_timestamp.cast(DataTypes.TimestampType), "Asia/Kolkata"), "yyyy-MM-dd'T'HH:mm:ss'Z'"))
   }
